iommu/qcom + drm/msm: correct fault handler return
The logic was inverted in qcom_iommu.  Although the only place a fault
handler was registered for this iommu driver is drm/msm.  Fix both
places at once to keep the behavior bisectable.

(drm/msm can also be used with with arm-smmu or msm_iommu, neither of
which report faults yet.)

Signed-off-by: Rob Clark <robdclark@chromium.org>