iommu/arm-smmu: implement stall/resume support
This is slightly over-complicated by the fact that not all hw can
reasonably support stall.  And the fact that we don't know which iommu
device a domain will be associated with when it is allocated.

This enables stall just for the QCOM implementation.  This might be
overly conservative but so far the only iommu user that *really* wants
stall is specific to qcom platforms.

Signed-off-by: Rob Clark <robdclark@chromium.org>