WIP: drm/msm: wire up fault state dump
We need a worker, as the fault callback will happen in an IRQ context
(at least for all current iommu drivers that drm/msm can possibly be
paired with).

But if the IOMMU supports stall/resume we can schedule a worker to
snapshot GPU state and dump cmdstream in a similar way to what we do
on GPU hangs.